// apps/factory/web/types/funding.ts
export interface DAOPool {
  daoId: string
  token: Address
  totalAccumulated: bigint
  contributorPool: bigint
  dependencyPool: bigint
  reservePool: bigint
  lastDistributedEpoch: number
  epochStartTime: number
}

export interface FundingEpoch {
  epochId: number
  daoId: string
  startTime: number
  endTime: number
  totalContributorRewards: bigint
  totalDependencyRewards: bigint
  totalDistributed: bigint
  finalized: boolean
}

export interface ContributorShare {
  contributorId: string
  weight: number
  pendingRewards: bigint
  claimedRewards: bigint
  lastClaimEpoch: number
}

export interface DependencyShare {
  depHash: string
  contributorId: string
  weight: number
  transitiveDepth: number
  usageCount: number
  pendingRewards: bigint
  claimedRewards: bigint
  isRegistered: boolean
}

export interface FeeDistributionConfig {
  treasuryBps: number
  contributorPoolBps: number
  dependencyPoolBps: number
  jejuBps: number
  burnBps: number
  reserveBps: number
}

export interface WeightVote {
  voter: Address
  targetId: string
  weightAdjustment: number
  reason: string
  reputation: number
  votedAt: number
}